fixes and tests
