+++ ts/test/test-utils/utils/message.ts
@@ -88,6 +88,7 @@ export class MockConversation {
       zombies: [],
       triggerNotificationsFor: 'all',
       isTrustedForAttachmentDownload: false,
+      isPinned: false,
     };
   }
